test(ui): split agent selector regression
